Update codahale/hdrhistogram import
This dependency has been moved to a new organization (now
`github.com/HdrHistogram/hdrhistogram`) and its use in this package
in `metrics/metricstest/local.go` under the old name broke
go-modules-based dependents, notably anyone using go modules
who depended on `jaegertracing/jaeger-client-go`.

This commit updates this one usage of the hdrhistogram dependency
and includes the relevant `dep` and `glide` updates.

The `glide.yaml` semver restriction on `prometheus/client_golang` was
adjusted from 2 down to 1.2, because any higher version has a
dependency on `github.com/cespare/xxhash/v2`, the `/v2` component of
which glide doesn't seem to be able to resolve. Prior to this change,
the version of `client_golang` recorded in the `glide.lock`
corresponded to v1.1.0, so this is not a regression.

Signed-off-by: Nathan Perry <np@nathanperry.dev>